refactor(backend): extract expert validator from handler
